extract required_* to its own module
