+++ wazo_amid/controller.py
@@ -6,6 +6,7 @@ import logging
 from threading import Thread
 from wazo_amid.ami.client import AMIClient
 from wazo_amid.bus.client import BusClient
+from wazo_amid import auth
 from wazo_amid import rest_api
 from wazo_amid.facade import EventHandlerFacade
 
@@ -40,14 +41,14 @@ class Controller:
 
     def _run_rest_api(self):
         self._token_renewer = TokenRenewer(AuthClient(**self._config['auth']))
+        rest_api.configure(self._config)
         if not rest_api.app.config['auth'].get('master_tenant_uuid'):
             self._token_renewer.subscribe_to_next_token_details_change(
-                rest_api.init_master_tenant
+                auth.init_master_tenant
             )
         self._token_renewer.subscribe_to_next_token_details_change(
             lambda t: self._token_renewer.emit_stop()
         )
-        rest_api.configure(self._config)
         with self._token_renewer:
             rest_api.run(self._config['rest_api'])
