+++ wazo_amid/rest_api.py
@@ -13,18 +13,12 @@ from flask_restful import Api
 from flask_restful import Resource
 from functools import wraps
 from werkzeug.contrib.fixers import ProxyFix
-from werkzeug.local import LocalProxy as Proxy
 from xivo import http_helpers
 from xivo import rest_api_helpers
 from xivo.http_helpers import ReverseProxied
-from xivo.auth_verifier import (
-    AuthVerifier,
-    required_acl as required_acl_,
-    required_tenant,
-)
+from xivo.auth_verifier import AuthVerifier
 
-
-from .exceptions import ValidationError, NotInitializedException
+from .exceptions import ValidationError
 
 VERSION = 1.0
 
@@ -33,7 +27,6 @@ logger = logging.getLogger(__name__)
 api = Api(prefix='/{}'.format(VERSION))
 auth_verifier = AuthVerifier()
 wsgi_server = None
-required_acl = required_acl_
 
 
 def configure(global_config):
@@ -100,28 +93,6 @@ def run(config):
 def stop():
     if wsgi_server:
         wsgi_server.stop()
-
-
-def init_master_tenant(token):
-    tenant_uuid = token['metadata']['tenant_uuid']
-    app.config['auth']['master_tenant_uuid'] = tenant_uuid
-
-
-def get_master_tenant_uuid(self):
-    if not app:
-        raise Exception('Flask application not configured')
-
-    tenant_uuid = app.config['auth'].get('master_tenant_uuid')
-    if not tenant_uuid:
-        raise NotInitializedException()
-    return tenant_uuid
-
-
-master_tenant_uuid = Proxy(get_master_tenant_uuid)
-
-
-def required_master_tenant():
-    return required_tenant(master_tenant_uuid)
 
 
 def handle_validation_exception(func):
